bin_tree: Add Search function and init_setup
Add an implementation to search if a value is present
in a binary tree, and also a helper function to setup an
initial binary tree (can help with some manipulations and testing)

The search function should also return the parent node.
This will help in subsequent operations like delete, which require
knowledge of parent.

Signed-off-by: Prashant Malani <pmalani@nvidia.com>