rb_tree: Fix delete implementations
Fix few errors:
In the case 4B of delete fixup, we were incorrectly
updating w->r to black, instead of w->l

Rotate helpfer functions did not take account of
the possibility that after rotation, the node might be
the new root, so modified them to update root if
required.

Lastly, added the -g option to the Makefile, so that
we can get line info while debugging using gdb.

Signed-off-by: Prashant Malani <p.malani@gmail.com>